requirements.txt
autograd==1.6.2
matplotlib==3.7.1
mosum==0.1.0
numpy==1.24.3
pandas==1.5.3
Pillow==10.2.0
psutil==5.9.0
rpy2==3.5.15
ruptures==1.1.8
scikit_learn==1.2.2
scipy==1.10.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
yfinance==0.2.36
